separate: app-specific common components (exclude from cloud)
