+++ static_report/src/components/Common/Footer.tsx
