+++ static_report/src/components/Common/Main.tsx
