+++ static_report/src/components/Common/Navbar.tsx
