+++ static_report/src/components/Common/Sidebar.tsx
