+++ static_report/src/components/Common/constant.ts
