+++ static_report/src/components/Tables/TableList/TableColumnSchemaList.tsx
@@ -16,7 +16,7 @@ import { FiChevronRight } from 'react-icons/fi';
 import { Comparable, Selectable } from '../../../types';
 import { NO_VALUE } from '../../Columns/constants';
 import { CompTableWithColEntryOverwrite } from '../../../utils/store';
-import { NO_DESCRIPTION_MSG } from '../../Layouts/constant';
+import { NO_DESCRIPTION_MSG } from '../../Common/constant';
 import { tableListWidth } from '../../../utils/layout';
 
 interface Props extends Selectable, Comparable {
