+++ static_report/src/components/Tables/TableList/TableListItem.tsx
@@ -21,7 +21,7 @@ import {
 } from '../../../utils/layout';
 import { ColumnSchemaDeltaSummary } from './ColumnSchemaDeltaSummary';
 import { CompTableColEntryItem, ReportState } from '../../../utils/store';
-import { NO_DESCRIPTION_MSG } from '../../Layouts/constant';
+import { NO_DESCRIPTION_MSG } from '../../Common/constant';
 import { getIconForColumnType } from '../../Columns';
 import { ColumnBadge } from './ColumnBadge';
